pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "muse-art"
version = "0.1.0"
description = "CLI art creation tool for engineers"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "openai>=1.0",
    "google-genai>=1.0",
    "term-image>=0.7",
    "tomli>=2.0;python_version<'3.12'",
    "tomli-w>=1.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-tmp-files>=0.0.2",
]

[project.scripts]
muse = "muse.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/muse"]

[tool.pytest.ini_options]
testpaths = ["tests"]
